Prevent router from republishing messages from WS
The router must be on a different channel than others, because of the
following: when a message is received from the platform it is published on
the bus.  If the router is on the same channels as others the message
event gets raised and the message is retransmitted to the platform.